+++ editor/src/components/canvas/commands/commands.ts
@@ -43,8 +43,8 @@ import type { DuplicateElement } from './duplicate-element-command'
 import { runDuplicateElement } from './duplicate-element-command'
 import type { UpdateFunctionCommand } from './update-function-command'
 import { runUpdateFunctionCommand } from './update-function-command'
-import type { PushIntendedBoundsAndUpdateTargets } from './push-intended-bounds-and-update-groups-command'
-import { runPushIntendedBoundsAndUpdateTargets } from './push-intended-bounds-and-update-groups-command'
+import type { PushIntendedBoundsAndUpdateTargets } from './push-intended-bounds-and-update-targets-command'
+import { runPushIntendedBoundsAndUpdateTargets } from './push-intended-bounds-and-update-targets-command'
 import type { DeleteProperties } from './delete-properties-command'
 import { runDeleteProperties } from './delete-properties-command'
 import type { AddImportsToFile } from './add-imports-to-file-command'
